+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### next
 
-* TODO: Replace this bullet point with an actual description of a change.
+* Switched to Zeitwerk as autoloader (#9)
 
 ### 3.7.0 (3 January 2025)
