+++ lib/conversejs/engine.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Conversejs
+  # The boilerplate Rails engine namespace.
+  class Engine < ::Rails::Engine; end
+end
