--- lib/conversejs/rails.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require 'conversejs/rails/engine'
-require 'conversejs/rails/version'
-
-# The root namespace of the conversjs gem.
-module Conversejs
-  # The Rails engine namespace.
-  module Rails; end
-end
