--- lib/conversejs/rails/engine.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Conversejs
-  module Rails
-    # The boilerplate Rails engine namespace.
-    class Engine < ::Rails::Engine; end
-  end
-end
